Rework swift-format invocation and data processing.
This effectively rewrites the core loop of the linter, enabling new
functionality and performing some micro-optimizations.

The use of `child_process.exec` is replaced with `child_process.spawn`.
This allows for the removal of the intermediatary shell process, which
is a micro-optimization in how the process executes.  This removal of
the shell allows this action to now be used on Windows as well as `exec`
would run the command in a subshell (i.e. `%COMSPEC% /c ...`) which
prevented us from reading the error messages.

The processing of the lint issue is changed to a regular expression.
The primary motivation for this is to handle the fact that using `:` as
a delimiter is problematic for Windows where `:` is the drive separator
in a path string, which is part of the tool's output.  We structure the
processing using the regular expression allowing us to ensure that we
have all the fields and properly handle paths on Windows.  By
destructuring the capture groups we also get a more readable reporting
format string.

We now invoke swift-format via the swift driver rather than directly
invoking the tool.  This is more in line with how the swift driver is
intended to be used to invoke custom subcommands.

Finally, we now report the number of issues detected since we must
manually track whether there was any output on stderr.  The incremental
cost is negligible and allows for a better reporting of the failure.