Extensive updates for the project management
The original author has stated that the project is not under active use and
development, and I had effectively forked the project.  This change brings
this project now to a point where it is barely a fork as there is so little
left in common.

Add a package.json to allow easy use of NPM.  This now explicitly lists the
dependencies of the project.

Take the opportunity to update the package dependencies to newer versions.
This update and explicit handling allows us to drop a number of installed
packages.

The GitHub APIs have changed, so take the opportunity to condense some of
the code during the update for the APIs.

The project is currently far less testable than before as the environment
variables to control the behaviour are gone and we no longer read the PR#
from a file.  It however avoids the unnecessary file system access.

The amount of changes that have been done now are extensive enough where I
feel it is proper to add (take?) the copyright by adding a notice for the
work.

Future avenues for improvements:
- drop additional configuration knobs (`excludes`, `excluded-types`)
- improve testing
- add infrastructure for packaging for distribution (esbuild)
- setup eslint
- setup GitHub workflows for testing/linting (who lint's the linter? :-p)